requirements.txt
antimony==2.14.0
appdirs==1.4.4
asgiref==3.8.1
asttokens==2.4.1
certifi==2024.2.2
charset-normalizer==3.3.2
comm==0.2.2
contourpy==1.2.1
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
Django==5.0.4
djangorestframework==3.15.1
executing==2.0.1
fonttools==4.51.0
idna==3.6
iniconfig==2.0.0
ipykernel==6.29.4
ipython==8.23.0
jedi==0.19.1
Jinja2==3.1.3
jupyter_client==8.6.1
jupyter_core==5.7.2
kiwisolver==1.4.5
libroadrunner==2.6.0
MarkupSafe==2.1.5
matplotlib==3.8.4
matplotlib-inline==0.1.6
nest-asyncio==1.6.0
numpy==1.26.4
packaging==24.0
pandas==2.2.1
parso==0.8.4
pexpect==4.9.0
phrasedml==1.3.0
pillow==10.3.0
platformdirs==4.2.0
plotly==5.20.0
pluggy==1.4.0
prompt-toolkit==3.0.43
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.17.2
pyparsing==3.1.2
pytest==8.1.1
python-dateutil==2.9.0.post0
python-libcombine==0.2.20
python-libnuml==1.1.6
python-libsbml==5.20.2
python-libsedml==2.0.32
pytz==2024.1
pyzmq==25.1.2
requests==2.31.0
scipy==1.13.0
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.3
tellurium==2.2.10
tenacity==8.2.3
tornado==6.4
traitlets==5.14.2
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
wcwidth==0.2.13
